Sprite editor
I have downloaded an image of Mario that I have then opened in the sprite editor, which was located on the window tab at the top and then opening the sprite editor, you can also open it up by using the image, then clicking it open. I then used all the buttons pointed out to me within the explanation to get an understanding of what they all do.