Add a secondary texture
I added in two more images of Mario and Luigi. However I have added the Mario as the secondary texture but I then I realised that this was for a different purpose like for a window or to add texture to a sprite not a sprite to a sprite. I then tried it with the example bus but I didn't have the original image so it didn't work. But I now do understand the purpose of this system.