9-slicing revisit
I have goon through the limitation The only two Collider2Ds that support 9-slicing are BoxCollider2D and PolygonCollider2D. You cannot edit BoxCollider2D or PolygonCollider2D when the Sprite Renderer’s Draw Mode is set to Sliced or Tiled. Editing in the Inspector window is disabled, and a warning appears to notify you that the Collider2D cannot be edited because it is driven by the Sprite Renderer component’s tiling properties. When the shape is regenerated in Auto Tiling, additional edges might appear within the Collider2D’s shape. This may have an effect on collisions. I tried these from the first page I then tried the different draw modes, Tiled, Sliced and Simple with simple not doing anything then inside of the tiling there was other adaptive being when you change the item the object will tile in a dynamic manor and sliced just taking the image and expanding the same texture out without tiling.